SW-4761 Create AccessionsRouter and move all accessions views and sub components to scenes directory
