--- src/components/accession2/index.tsx
@@ -1,3 +1,0 @@
-import Accession2View from './view/Accession2View';
-import Accession2Create from './create/Accession2Create';
-export { Accession2View, Accession2Create };
