+++ src/scenes/AccessionsRouter/Accession2View.tsx
@@ -9,30 +9,30 @@ import { Accession } from 'src/types/Accession';
 import { AccessionService, FacilityService } from 'src/services';
 import strings from 'src/strings';
 import TfMain from 'src/components/common/TfMain';
-import DeleteAccessionModal from '../edit/DeleteAccessionModal';
-import DetailPanel from './DetailPanel';
-import EditLocationModal from '../edit/EditLocationModal';
-import EditStateModal from '../edit/EditStateModal';
-import EndDryingReminderModal from '../edit/EndDryingReminderModal';
-import QuantityModal from '../edit/QuantityModal';
-import WithdrawModal from '../withdraw/WithdrawModal';
-import CheckedInConfirmationModal from '../edit/CheckedInConfirmationModal';
-import Accession2History from '../history/Accession2History';
-import ViabilityTestingPanel from '../viabilityTesting/ViabilityTestingPanel';
+import DeleteAccessionModal from './edit/DeleteAccessionModal';
+import DetailPanel from './view/DetailPanel';
+import EditLocationModal from './edit/EditLocationModal';
+import EditStateModal from './edit/EditStateModal';
+import EndDryingReminderModal from './edit/EndDryingReminderModal';
+import QuantityModal from './edit/QuantityModal';
+import WithdrawModal from './withdraw/WithdrawModal';
+import CheckedInConfirmationModal from './edit/CheckedInConfirmationModal';
+import Accession2History from './history/Accession2History';
+import ViabilityTestingPanel from './viabilityTesting/ViabilityTestingPanel';
 import useSnackbar from 'src/utils/useSnackbar';
 import useQuery from 'src/utils/useQuery';
 import useStateLocation, { getLocation } from 'src/utils/useStateLocation';
 import useDeviceInfo from 'src/utils/useDeviceInfo';
 import PageSnackbar from 'src/components/PageSnackbar';
-import ViabilityModal from '../edit/ViabilityModal';
-import NewViabilityTestModal from '../viabilityTesting/NewViabilityTestModal';
+import ViabilityModal from './edit/ViabilityModal';
+import NewViabilityTestModal from './viabilityTesting/NewViabilityTestModal';
 import { ViabilityTest } from 'src/types/Accession';
-import ViewViabilityTestModal from '../viabilityTesting/ViewViabilityTestModal';
+import ViewViabilityTestModal from './viabilityTesting/ViewViabilityTestModal';
 import { getSeedBank, isContributor } from 'src/utils/organization';
 import _ from 'lodash';
-import PageHeaderWrapper from '../../common/PageHeaderWrapper';
+import PageHeaderWrapper from '../../components/common/PageHeaderWrapper';
 import { APP_PATHS } from 'src/constants';
-import OverviewItemCard from '../../common/OverviewItemCard';
+import OverviewItemCard from '../../components/common/OverviewItemCard';
 import BackToLink from 'src/components/common/BackToLink';
 import { useLocalization, useUser } from 'src/providers';
 import { useOrganization } from 'src/providers/hooks';
