+++ src/scenes/AccessionsRouter/view/DetailPanel.tsx
