+++ src/scenes/AccessionsRouter/withdraw/WeightWithdrawal.tsx
